code refactor - splitting views into different files
